maybe

+++ server.js
@@ -16,18 +16,19 @@ let baseDir;
 let chokidarTimeout;
 let oldBackendTemp;
 let removeBackendFolderInterval;
+let removeFilePath;
 
 export default async function createServer(type = 'dev') {
   try {
     const startTime = Date.now();
     baseDir = import.meta.dirname.split('node_modules')[0];
+    removeFilePath = path.join(baseDir, 'backend', '__remove_me.txt');
 
     if (!removeBackendFolderInterval) {
-      const removeMeFilePath = path.join(baseDir, 'backend', '__remove_me.txt');
       removeBackendFolderInterval = setInterval(() => {
 
-        if (fs.existsSync(removeMeFilePath)) {
-          fs.rmSync(removeMeFilePath);
+        if (fs.existsSync(removeFilePath)) {
+          fs.rmSync(removeFilePath);
           try {
             globalThis.openDbFromQueryMaker && globalThis.openDbFromQueryMaker.close();
           }
@@ -139,7 +140,7 @@ async function addBackend(app) {
   }
 
   // copy the whole backend to a temp folder - since we do not want cached imports
-  if (fs.existsSync(backendFolder)) {
+  if (fs.existsSync(backendFolder) && !fs.existsSync(removeFilePath)) {
     let tempFolder = path.join(import.meta.dirname, 'tempBackends', Date.now() + '');
     fs.cpSync(backendFolder, tempFolder, { recursive: true });
